+++ Sources/SwifterSwift/SwiftStdlib/Deprecated/StdlibDeprecated.swift
@@ -79,3 +79,33 @@ public extension Array {
     }
 
 }
+
+public extension Sequence {
+
+    /// SwifterSwift: Returns an array containing the results of mapping the given key path over the sequence’s elements.
+    ///
+    /// - Parameter keyPath: Key path to map.
+    /// - Returns: An array containing the results of mapping.
+    @available(*, deprecated, message: "Please use map() with a key path instead.")
+    func map<T>(by keyPath: KeyPath<Element, T>) -> [T] {
+        return map { $0[keyPath: keyPath] }
+    }
+
+    /// SwifterSwift: Returns an array containing the non-nil results of mapping the given key path over the sequence’s elements.
+    ///
+    /// - Parameter keyPath: Key path to map.
+    /// - Returns: An array containing the non-nil results of mapping.
+    @available(*, deprecated, message: "Please use compactMap() with a key path instead.")
+    func compactMap<T>(by keyPath: KeyPath<Element, T?>) -> [T] {
+        return compactMap { $0[keyPath: keyPath] }
+    }
+
+    /// SwifterSwift: Returns an array containing the results of filtering the sequence’s elements by a boolean key path.
+    ///
+    /// - Parameter keyPath: Boolean key path. If it's value is `true` the element will be added to result.
+    /// - Returns: An array containing filtered elements.
+    @available(*, deprecated, message: "Please use filter() with a key path instead.")
+    func filter(by keyPath: KeyPath<Element, Bool>) -> [Element] {
+        return filter { $0[keyPath: keyPath] }
+    }
+}